Fix FTBFS on ARM introduced in PR #103
PR #103 tried to fix a linking issue in Visual Studio 2013 on
Windows 7. It added multiple explicit template instantiations
which violates the standard and results in failure to build
from source on ARM.

Further testing failed to reproduce the linking issue with
Visual Studio 2013 on Windows 8.1. Thus this commit removes
the explicit template instantiations.